A working buildout recipe for setiing up and using virtualenv
This is an improvement on what I've seen others attempt for buildout,
but not quiet achieve to my satisfaction.  I've had to reimplement
methods from virtualenv myself, as if I use it as is, it will fail to
implement a virtualenv.  This is all to do with buildout using it's own
copy of site.py, which virtualenv relies on to figure out where system
packages are, etc.

This was a big learning curve for me, and had me cursing both buildout
and virtualenv designers, for the way they implemented things.